requirements.txt
openai==0.28
pyyaml
python-dotenv
tqdm
